% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_source_server.R
\name{annotate_source_server}
\alias{annotate_source_server}
\title{Load course data}
\usage{
annotate_source_server(id, documents, references, study_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{documents}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{references}{Reactive. List of references.}

\item{study_paths}{Reactive. Paths to project databases and sub-folders.}
}
\value{
A list of course data.
}
\description{
Module facilitating the loading of all relevant course data.
}
\author{
Nicolas Mangin
}
