% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{anova_test}
\alias{anova_test}
\title{Analysis of Variance /ANOVA/ Test}
\usage{
anova_test(data, x, y)
}
\arguments{
\item{data}{This is the data set containing the variables you want use.}

\item{x}{Predictor variable.}

\item{y}{Outcome variable.}
}
\value{
Some summary data related to the test:
means of each group, p-value, t-statistic.
}
\description{
Conducts a Welch's ANOVA.
Best to use when your predictor variable is categorical,
and your outcome variable is quantitative.
}
