% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar plot Builder}
\usage{
bar_plot(
  data,
  x,
  y = NULL,
  title = "I don't like nice titles",
  subtitle = "And I HATE good subtitles",
  x_name = rlang::as_label(rlang::enquo(x)),
  y_name = rlang::as_label(rlang::enquo(y)),
  flip = FALSE
)
}
\arguments{
\item{data}{This is the data set containing the variables you want to visualize.}

\item{x}{Variable you want to see the count of by level. Ex: Number of brown squirrels vs. gray squirrels vs. black squirrels. The x-variable here would be "fur color". Must be categorical}

\item{y}{Another variable you could use to observe proportions. Ex: Number of brown squirrels that are from Brooklyn vs. Queens vs. Manhattan. The y-variable here would be "borough". Must be categorical.}

\item{title}{A string containing the title of your graph}

\item{subtitle}{A string containing the subtitle of your graph}

\item{x_name}{The name of your x-variable. Ex: your variable might be `b_birds`, but you may want it to appear as "Brown Birds" on the graph.}

\item{y_name}{The name of your y-variable.}

\item{flip}{Set to TRUE or FALSE if you would like to flip the plot's axes.}
}
\value{
A bar plot made to the user's specifications
}
\description{
Builds a basic bar plot with the ability to add a second dimension.
}
