% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{box_plot}
\alias{box_plot}
\title{Box plot Builder}
\usage{
box_plot(
  data,
  x,
  y = NULL,
  title = "I don't like nice titles",
  subtitle = "And I HATE good subtitles",
  x_name = rlang::as_label(rlang::enquo(x)),
  y_name = rlang::as_label(rlang::enquo(y))
)
}
\arguments{
\item{data}{This is the data set containing the variables you want to visualize.}

\item{x}{Variable you want to visualize the distribution of. This must be a numeric variable.}

\item{y}{Another variable you could use to observe the distribution of your original variable by level. Ex: Heights of squirrels by borough—here, "borough" would be the y variable. This variable must be categorical.}

\item{title}{A string containing the title of your graph}

\item{subtitle}{A string containing the subtitle of your graph}

\item{x_name}{The name of your x-variable. Ex: your variable might be `b_birds`, but you may want it to appear as "Brown Birds" on the graph.}

\item{y_name}{The name of your y-variable.}
}
\value{
A box plot made to the user's specifications
}
\description{
Builds a basic box plot with the ability to add a second dimension.
}
