% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_variables.R
\name{check_missing_variables}
\alias{check_missing_variables}
\title{Checking a dataset for missing observations across variables}
\usage{
check_missing_variables(dataset)
}
\arguments{
\item{dataset}{A vector, matrix, data frame, or tibble.}
}
\value{
If there is no missing data the function will return a string
informing you that there is no missing data. Otherwise a graph will
appear showing the variable one the y axis and the percent missing on the x-axis.
}
\description{
This function returns a graph that shows variable missingness across the
entire dataset.
}
\examples{
check_missing_variables(cars) # no missing data
check_missing_variables(airquality) # shows missing data
}
