% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{clean_text}
\alias{clean_text}
\title{Text Cleaner}
\usage{
clean_text(txt)
}
\arguments{
\item{txt}{The string of text you would like cleaned.}
}
\value{
A string that is the cleaned version of the one provided.
}
\description{
Prepares text data for most analyses.
}
