% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{column_plot}
\alias{column_plot}
\title{Column plot Builder}
\usage{
column_plot(
  data,
  x,
  y,
  color_by = NULL,
  title = "I don't like nice titles",
  subtitle = "And I HATE good subtitles",
  x_name = rlang::as_label(rlang::enquo(x)),
  y_name = rlang::as_label(rlang::enquo(y)),
  color_name = rlang::as_label(rlang::enquo(color_by)),
  flip = FALSE
)
}
\arguments{
\item{data}{This is the data set containing the variables you want to visualize.}

\item{x}{Variable for your x-axis. Must be categorical}

\item{y}{Variable for your y-axis. Must be numeric.}

\item{color_by}{Another variable you can use to observe proportion. Must be categorical}

\item{title}{A string containing the title of your graph}

\item{subtitle}{A string containing the subtitle of your graph}

\item{x_name}{The name of your x-variable. Ex: your variable might be `b_birds`, but you may want it to appear as "Brown Birds" on the graph.}

\item{y_name}{The name of your y-variable.}

\item{color_name}{The name of your color_by variable.}

\item{flip}{Set to TRUE or FALSE if you would like to flip the plot's axes.}
}
\value{
A column plot made to the user's specifications
}
\description{
Builds a basic column plot with the ability to add a third dimension.
}
