% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_predictions.R
\name{create_predictions}
\alias{create_predictions}
\title{Creating predictions using simulated data}
\usage{
create_predictions(model, dataset, ..., model_type = "ols")
}
\arguments{
\item{model}{The model used to create the predictiosn}

\item{dataset}{The underlying dataset that is used to create the underlying
values of the predictor variable}

\item{...}{List of predictor variables that are being varied.  All other variables are set to 'typical' values}

\item{model_type}{Specifies whether you have an 'ols' or 'logit' model}
}
\description{
This function returns a data frame of predictions using simulated data.
This result only works on OLS or logit models.
}
