% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document_types.R
\docType{data}
\name{data_document_types}
\alias{data_document_types}
\title{Document types}
\format{
\itemize{
\item type: factor. Possible document types: Note,Page,Slide,Video, Game, Statements,Alternatives,Computation, Essay, Problem, or Case.
\item icon: character. Font awesome icon associated to the document type.
\item boxcolor: character. Box color associated to the document type.
\item description: character. Description of the type of document.
}
}
\usage{
data("data_document_types")
}
\description{
Types of documents which can be created and associated markers.
}
\author{
Nicolas Mangin
}
\keyword{types}
