% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_folders_structure.R
\docType{data}
\name{data_folders_structure}
\alias{data_folders_structure}
\title{Course folder structure}
\format{
Tibble with
\itemize{
\item folder: character. Name of the main the folder.
\item path: character. Path to the specific folder.
}
}
\usage{
data("data_folders_structure")
}
\description{
List of folder paths to generate a course project
}
\author{
Nicolas Mangin
}
\keyword{course}
\keyword{structure}
