% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_references.R
\docType{data}
\name{data_references}
\alias{data_references}
\title{References}
\format{
\itemize{
\item key: character. Identifier of the references.
\item order: character. Number of the entry.
\item bibtype: factor. Bibtypes: Article, InCollection, Book, Unpublished, InProceedings, Online, Review, Letter, Misc, Booklet, InBook, Manual, MasterThesis, PhdThesis, Proceedings.
\item author: character. Authors of the references. Formatted as Doe, Jane and Doe, John.
\item title: character. Title of the reference.
\item journal: character. Journal in which the article was published (if this is an article).
\item jnl: character. Acronym of the journal in which the article was published (if this is an article).
\item issn: character. ISSN of the journal in which the article was published (if this is an article).
\item field: character. Discipline in which the article was published (if this is an article).
\item year: integer. Year of publication.
\item volume: integer. Volume in which the article was published (if this is an article).
\item number: integer. Issue in which the article was published (if this is an article).
\item pages: character. Pages on which the document was published (if this is an article or book section).
\item doi: character. DOI of the publication.
\item abstract: character. Abstract.
\item keywords: character. Keywords.
\item url: character. URL (web link) to the document.
\item publisher: character. Publisher of the book.
\item booktitle: character. Title of the book in which the book section or chapter was published.
\item editor: character. Editor of the book.
\item address: character. Address of publication (for books).
\item chapter: character. Chapter number of the section.
\item edition: character. Number of the edition.
\item isbn: character. ISBM of the book.
\item comment: character. Comments on the document.
\item note: character. Note on the document.
}
}
\usage{
data("data_references")
}
\description{
List of references with all the field necessary to export as a bib file.
}
\author{
Nicolas Mangin
}
\keyword{bib}
\keyword{latex}
\keyword{references}
