% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_return.R
\name{data_return}
\alias{data_return}
\title{data_return}
\usage{
data_return(x)
}
\arguments{
\item{x}{a numeric vector containing a positive integer (year) between 1985 - 2020}
}
\value{
A dataframe of funded NIH research opportunities relevant to the year input (x)
}
\description{
data_return
}
