% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tags.R
\docType{data}
\name{data_tags}
\alias{data_tags}
\title{Tags}
\format{
\itemize{
\item tag: character. Label of the tag.
\item order: integer. Order of the tag category.
\item value: character. Categories within the tag.
\item count: integer. Number of documents associated to this tag.
\item icon: character. Font awesome icon associated to the tag.
\item boxcolor: character. Box color associated to the the tag.
}
}
\usage{
data("data_tags")
}
\description{
Description of tags ans associated markers.
}
\author{
Nicolas Mangin
}
\keyword{classification}
\keyword{document}
\keyword{tags}
