% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{density_plot}
\alias{density_plot}
\title{Density plot Builder}
\usage{
density_plot(
  data,
  x,
  color_by = NULL,
  facet_by = NULL,
  title = "I don't like nice titles",
  subtitle = "And I HATE good subtitles",
  x_name = rlang::as_label(rlang::enquo(x)),
  color_name = rlang::as_label(rlang::enquo(color_by))
)

density_plot(
  data,
  x,
  color_by = NULL,
  facet_by = NULL,
  title = "I don't like nice titles",
  subtitle = "And I HATE good subtitles",
  x_name = rlang::as_label(rlang::enquo(x)),
  color_name = rlang::as_label(rlang::enquo(color_by))
)
}
\arguments{
\item{data}{This is the data set containing the variables you want to visualize.}

\item{x}{Variable you want to visualize the distribution of. This must be a numeric variable.}

\item{color_by}{Another variable you could use to observe the distribution of your original variable by level. Ex: Heights of squirrels by borough—here, "borough" would be the color_by variable. This variable must be categorical.}

\item{facet_by}{Another variable you could use to observe the distribution of your original variable by level. Must be categorical.}

\item{title}{A string containing the title of your graph}

\item{subtitle}{A string containing the subtitle of your graph}

\item{x_name}{The name of your x-variable. Ex: your variable might be `b_birds`, but you may want it to appear as "Brown Birds" on the graph.}

\item{color_name}{The name of your color variable.}
}
\value{
A density plot made to the user's specifications

A density plot made to the user's specifications
}
\description{
Builds a basic density plot with the ability to add a second dimension.

Builds a basic density plot with the ability to add a second dimension.
}
