% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_draw.R
\name{diagram_draw}
\alias{diagram_draw}
\title{Create a diagram}
\usage{
diagram_draw(nodes, relations, moderations)
}
\arguments{
\item{nodes}{Tibble. Nodes with their properties.}

\item{relations}{Tibble. Edges with their properties.}

\item{moderations}{Tibble. Moderations with their properties.}
}
\value{
A DiagrammeR object ready for rendering.
}
\description{
Function creating a diagram object from tibbles specifying nodes, relations, and moderations.
}
\author{
Nicolas Mangin
}
