% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_edit_server.R, R/diagram_edit_ui.R
\name{diagram_edit_ui}
\alias{diagram_edit_ui}
\alias{diagram_edit_server}
\title{Edit diagrams}
\usage{
diagram_edit_server(id)

diagram_edit_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Write lines of codes creating the different parts of a diagram which can then be embedded in function or document.

Write lines of codes creating the different parts of a diagram which can then be embedded in function or document.
}
\description{
Module facilitating the quick creation of diagrams embedded in functions or documents.

Module facilitating the quick creation of diagrams embedded in functions or documents.
}
\author{
Nicolas Mangin

Nicolas Mangin
}
