% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_initialize.R
\name{diagram_initialize}
\alias{diagram_initialize}
\title{Create a starting diagram}
\usage{
diagram_initialize()
}
\value{
A list of tibbles specifying nodes, relations, and moderations.
}
\description{
Function creating a basic diagram serving as a basis to be adjusted.
}
\author{
Nicolas Mangin
}
