% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_update.R
\name{diagram_update}
\alias{diagram_update}
\title{Update a diagram}
\usage{
diagram_update(nodes, relations, moderations)
}
\arguments{
\item{nodes}{Tibble. Nodes with their properties.}

\item{relations}{Tibble. Edges with their properties.}

\item{moderations}{Tibble. Moderations with their properties.}
}
\value{
A list of tibbles specifying nodes, relations, and moderations.
}
\description{
Function updating a diagram serving based on user input.
}
\author{
Nicolas Mangin
}
