% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_write_table_code.R
\name{diagram_write_table_code}
\alias{diagram_write_table_code}
\title{Write a table as code}
\usage{
diagram_write_table_code(table, name)
}
\arguments{
\item{table}{Data.frame or tibble. Table which should be written as code.}

\item{name}{Character. Name of the resulting table.}
}
\value{
A vector of character in which entries are lines to be written.
}
\description{
Take a table and transforms it into lines of code which can be written and inserted.
}
\author{
Nicolas Mangin
}
