% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents_edit_ui.R
\name{documents_edit_ui}
\alias{documents_edit_ui}
\title{Edit notes on the literature}
\usage{
documents_edit_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified note in the folder "3_documents".
}
\description{
Module facilitating the quick edition of notes on the literature.
}
\author{
Nicolas Mangin
}
