% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents_list.R
\name{documents_list}
\alias{documents_list}
\title{List all document}
\usage{
documents_list()
}
\value{
Save the document list as a tibble in the folder "2_documents" and returns this list
}
\description{
Function listing all the documents, checking whether they were changed, and update information about changed documents.
}
\author{
Nicolas Mangin
}
