% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents_make_new_name.R
\name{documents_make_new_name}
\alias{documents_make_new_name}
\title{Create a document name}
\usage{
documents_make_new_name(firstletter)
}
\arguments{
\item{firstletter}{Character. "L" for literature, "N" for note, "A" for article, "T" for Thesis, "P" for presentation, "V" for video, and "F" for feedback.}
}
\value{
Character. Name of a new document to be created.
}
\description{
Function listing the names of the existing documents and making a new name by adding one to the highest value. This ensures that no old number is assigned again.
}
\author{
Nicolas Mangin
}
