% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_results.R
\name{explain_results}
\alias{explain_results}
\title{Explaining Results for OLS models}
\usage{
explain_results(model_name, variable)
}
\arguments{
\item{model_name}{The oLS model created}

\item{variable}{The independent variable you want explained}
}
\description{
This function returns a string for a specific variable that explains in simple terms
the influence of the independent variable on the dependent variable.
}
\examples{
model1 <- lm(mpg ~ wt, data = mtcars)
model2 <- lm(mpg ~ wt + cyl + hp + wt + disp, data = mtcars)
explain_results(model1, "wt")
explain_results(model2, "cyl")
}
