% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_bivariate.R
\name{explore_bivariate}
\alias{explore_bivariate}
\title{Exploring biviate regression results of a dataframe}
\usage{
explore_bivariate(data, dependent, independent, p_value = 0.05,
  type = "graph", model_type = "ols")
}
\arguments{
\item{data}{A matrix, data frame, or tibble.}

\item{dependent}{A string character of the dependent variable}

\item{independent}{A string vector of the different independent variables to test}

\item{p_value}{The significance threshold for a table output (defaults to 0.05). Graph is
fixed at 95 percent confidence intervals}

\item{type}{Specifies the type of output request.  either 'graph' or 'table'}

\item{model_type}{Specificies whether you have an OLS or logit model.}
}
\description{
This function returns either a graph or table of bivariate regression results for a dataset.
The results are helpful when trying to first gain an understanding of relationships between
a large number of variables.
}
\examples{
explore_bivariate(gapminder::gapminder, "lifeExp", c("gdpPercap", "continent"), type = "table")
explore_bivariate(gapminder::gapminder, "lifeExp", c("gdpPercap", "continent"))
explore_bivariate(mtcars, "mpg", c("cyl", "disp", "hp", "drat"), p_value = 0.00001)
explore_bivariate(mtcars, names(mtcars)[1], names(mtcars)[-1])
explore_bivariate(titanic::titanic_train, "Survived", c("Sex", "Age"), model_type = "logit")
}
