% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_edit_server.R
\name{functions_edit_server}
\alias{functions_edit_server}
\title{Edit functions}
\usage{
functions_edit_server(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified function in the folder "2_functions".
}
\description{
Module facilitating the quick creation or modification of functions used in documents.
}
\author{
Nicolas Mangin
}
