% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_edit_ui.R
\name{functions_edit_ui}
\alias{functions_edit_ui}
\title{Edit functions}
\usage{
functions_edit_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified function in the folder "1_preparation/functions".
}
\description{
Module facilitating the quick creation or modification of functions used in documents.
}
\author{
Nicolas Mangin
}
