% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding_frequency.R
\name{funding_frequency}
\alias{funding_frequency}
\title{funding_frequency}
\usage{
funding_frequency(data_frame)
}
\arguments{
\item{data_frame}{data of NIH research opportunities for year selected by user}
}
\value{
an interactive graph of funding frequency by state
}
\description{
funding_frequency
}
