% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{line_plot}
\alias{line_plot}
\title{Line Plot Builder}
\usage{
line_plot(
  data,
  x,
  y,
  color_by = NULL,
  line_type = NULL,
  title = "I don't like nice titles",
  subtitle = "And I HATE good subtitles",
  x_name = rlang::as_label(rlang::enquo(x)),
  y_name = rlang::as_label(rlang::enquo(y)),
  color_name = rlang::as_label(rlang::enquo(color_by)),
  line_name = rlang::as_label(rlang::enquo(line_type))
)
}
\arguments{
\item{data}{This is the data set containing the variables you want to visualize.}

\item{x}{Variable for your x-axis. This must be numeric AND continuous. Ex: Time is often a variable used here.}

\item{y}{Variable for your y-axis. This must be numeric.}

\item{color_by}{Another variable you could use to conditionally change the color of your points. Must be categorical.}

\item{line_type}{Another variable you could use to conditionally change the type of your line. This will only work with a categorical variable.}

\item{title}{A string containing the title of your graph}

\item{subtitle}{A string containing the subtitle of your graph}

\item{x_name}{The name of your x-variable. Ex: your variable might be `b_birds`, but you may want it to appear as "Brown Birds" on the graph.}

\item{y_name}{The name of your y-variable.}

\item{color_name}{The name of your color_by variable.}

\item{line_name}{The name of your size_by variable.}
}
\value{
A line plot made to the user's specifications
}
\description{
Builds a basic line plot with the ability to add a third dimension and fourth dimension.
}
