% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature_edit_server.R
\name{literature_edit_server}
\alias{literature_edit_server}
\title{Edit notes on the literature}
\usage{
literature_edit_server(id, selected_documents)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{selected_documents}{Tibble. List of documents.}
}
\value{
Save the new or modified note in the folder "3_documents".
}
\description{
Module facilitating the quick edition of notes on the literature.
}
\author{
Nicolas Mangin
}
