% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_study_server.R
\name{load_study_server}
\alias{load_study_server}
\title{Load course data}
\usage{
load_study_server(id, study_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{study_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
A list of course data.
}
\description{
Module facilitating the loading of all relevant course data.
}
\author{
Nicolas Mangin
}
