% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_study_ui.R
\name{load_study_ui}
\alias{load_study_ui}
\title{Load course data}
\usage{
load_study_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list of course data.
}
\description{
Module facilitating the loading of all relevant course data.
}
\author{
Nicolas Mangin
}
