% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic Regression}
\usage{
logistic_regression(data, x, y)
}
\arguments{
\item{data}{This is the data set containing the variables you want use.}

\item{x}{Predictor variable.}

\item{y}{Outcome variable.}
}
\value{
Some summary data related to the regression:
regression coefficient, p-value, Adjusted R-Squared.
}
\description{
Creates a logistic regression model.
Best to use when your predictor variable is quantitative,
and your outcome variable is binary.
}
