% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_instruments_server.R
\name{make_instruments_server}
\alias{make_instruments_server}
\title{Make survey instruments}
\usage{
make_instruments_server(id, documents, study_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{documents}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{study_paths}{Reactive. Paths to project databases and sub-folders.}

\item{references}{Reactive. List of references.}
}
\value{
Modify instruments and databases
}
\description{
Module facilitating the creation of question, questionnaires, and experiments.
}
\author{
Nicolas Mangin
}
