% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_instruments_ui.R
\name{make_instruments_ui}
\alias{make_instruments_ui}
\alias{annotate_source_ui}
\title{Make survey instruments}
\usage{
annotate_source_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Modify instruments and databases
}
\description{
Module facilitating the creation of question, questionnaires, and experiments.
}
\author{
Nicolas Mangin
}
