% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_id.R
\name{make_new_id}
\alias{make_new_id}
\title{New id for databases}
\usage{
make_new_id(x, y)
}
\arguments{
\item{x}{Character vector. Vector of existing ids.}

\item{y}{Character. Prefix of the new id.}
}
\value{
Character. New id
}
\description{
Create new ids by incrementing existing ones.
}
\author{
Nicolas Mangin
}
