% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_total_cost.R
\name{median_total_cost}
\alias{median_total_cost}
\title{median_total_cost}
\usage{
median_total_cost(your_dataFrame)
}
\arguments{
\item{your_dataFrame}{an nih exporter data frame extracted by the nih_research function}
}
\value{
An interactive bar graph: the median total cost v.s the number of support year
}
\description{
median_total_cost
}
