% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_research.R
\name{nih_research}
\alias{nih_research}
\title{nih_research}
\usage{
nih_research(year)
}
\arguments{
\item{year}{a numeric vector containing a positive integer (year) between 1985 - 2020}
}
\value{
A dataframe of funded NIH research opportunities relevant to the year input (year)
}
\description{
nih_research
}
