% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_edit_server.R
\name{processing_edit_server}
\alias{processing_edit_server}
\title{Edit script and update objects}
\usage{
processing_edit_server(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified script in the root folder.
}
\description{
Module facilitating the quick edition of script and update objects for targets
}
\author{
Nicolas Mangin
}
