% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_filter_server.R
\name{references_filter_server}
\alias{references_filter_server}
\title{Filter a list of references}
\usage{
references_filter_server(id, references)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{references}{Tibble. List of references.}
}
\value{
Filter list of references
}
\description{
Mudule allowing the user to find references using various filters.
}
\author{
Nicolas Mangin
}
