% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_filter_ui.R
\name{references_filter_ui}
\alias{references_filter_ui}
\title{Filter a list of references}
\usage{
references_filter_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Filter list of references
}
\description{
Module allowing the user to find references using various filters.
}
\author{
Nicolas Mangin
}
