% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_make_bib_entry.R
\name{references_make_bib_entry}
\alias{references_make_bib_entry}
\title{Write an entry for a bibtex file}
\usage{
references_make_bib_entry(
  bibtype,
  key,
  author,
  year,
  title,
  journal,
  volume,
  number,
  pages,
  doi,
  url,
  publisher,
  booktitle,
  editor,
  address,
  chapter,
  edition,
  isbn
)
}
\arguments{
\item{bibtype}{Character.}

\item{key}{Character.}

\item{author}{Character.}

\item{year}{Character.}

\item{title}{Character.}

\item{journal}{Character.}

\item{volume}{Character.}

\item{number}{Character.}

\item{pages}{Character.}

\item{doi}{Character.}

\item{url}{Character.}

\item{publisher}{Character.}

\item{booktitle}{Character.}

\item{editor}{Character.}

\item{address}{Character.}

\item{chapter}{Character.}

\item{edition}{Character.}

\item{isbn}{Character.}
}
\value{
A character vector in which each elements is a row of a bib entry.
}
\description{
Function taking one row of a reference list and writing the corresponding entry.
}
\author{
Nicolas Mangin
}
