% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_with_childrow.R
\name{references_with_childrow}
\alias{references_with_childrow}
\title{Table with collapsible rows (1)}
\usage{
references_with_childrow(x, vars = NULL, opts = NULL, ...)
}
\arguments{
\item{x}{Data.frame. Table of references.}

\item{vars}{Character vector. Name of the variables to collapse.}

\item{opts}{List. List of options.}

\item{...}{Other arguments.}
}
\value{
A data.table with collapsed rows.
}
\description{
Function formatting references with collapsible rows
}
\author{
Nicolas Mangin
}
