% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchr.R
\docType{package}
\name{researchr-package}
\alias{researchr}
\alias{researchr-package}
\title{researchr: Automating AccessLex Analysis}
\description{
Automating everything from exploration, charts, and checking assumptions
}
\author{
\strong{Maintainer}: Richard Gardiner \email{rgardiner90@gmail.com}

}
