% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter Plot Builder}
\usage{
scatter_plot(
  data,
  x,
  y,
  color_by = NULL,
  size_by = NULL,
  title = "I don't like nice titles",
  subtitle = "And I HATE good subtitles",
  x_name = rlang::as_label(rlang::enquo(x)),
  y_name = rlang::as_label(rlang::enquo(y)),
  color_name = rlang::as_label(rlang::enquo(color_by)),
  size_name = rlang::as_label(rlang::enquo(size_by)),
  line = TRUE,
  se = FALSE
)
}
\arguments{
\item{data}{This is the data set containing the variables you want to visualize.}

\item{x}{Variable for your x-axis}

\item{y}{Variable for your y-axis}

\item{color_by}{Another variable you could use to conditionally change the color of your points. If numeric, the variable will manifest as a color gradient—if categorical the variable will manifest as a discrete set of colors.}

\item{size_by}{Another variable you could use to conditionally change the size of your points. This will only work with a numeric variable.}

\item{title}{A string containing the title of your graph}

\item{subtitle}{A string containing the subtitle of your graph}

\item{x_name}{The name of your x-variable. Ex: your variable might be `b_birds`, but you may want it to appear as "Brown Birds" on the graph.}

\item{y_name}{The name of your y-variable.}

\item{color_name}{The name of your color_by variable.}

\item{size_name}{The name of your size_by variable.}

\item{line}{Set to TRUE or FALSE if you would or would not like a line of best fit drawn through your points.}

\item{se}{Set to TRUE or FALSE if you would or would not like the standard error visualized with your line of best fit.}
}
\value{
A scatter plot made to the user's specifications
}
\description{
Builds a basic scatter plot with the ability to add a third and fourth dimension.
}
