% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{simple_linear_regression}
\alias{simple_linear_regression}
\title{Simple Linear Regression}
\usage{
simple_linear_regression(data, x, y)
}
\arguments{
\item{data}{This is the data set containing the variables you want use.}

\item{x}{Predictor variable.}

\item{y}{Outcome variable.}
}
\value{
Some summary data related to the regression:
regression coefficient, p-value, Adjusted R-Squared.
}
\description{
Creates a simple linear regression, or SLR, model.
Best to use when your predictor AND outcome variable are quantitative.
}
