% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{t_test}
\alias{t_test}
\title{T-Test}
\usage{
t_test(data, x, y)
}
\arguments{
\item{data}{This is the data set containing the variables you want use.}

\item{x}{Predictor variable.}

\item{y}{Outcome variable.}
}
\value{
Some summary data related to the test:
means of each group, p-value, t-statistic.
}
\description{
Conducts a Welch's T-Test.
Best to use when your predictor variable is binary,
and your outcome variable is quantitative.
}
